requirements.txt
gitPython
matplotlib
boto3
psutil
PyYAML

